Updates to repo structure, setup file and adding binary
