+++ build.sh
@@ -1,0 +1,2 @@
+pip3 install -r requirements.txt
+pyinstaller --onefile src/tfvm.py
