+++ src/requirements.txt
@@ -1,0 +1,11 @@
+blessed==1.17.6
+readchar==2.0.1
+python-editor==1.0.4
+inquirer==2.7.0
+requests==2.25.1
+certifi==2020.12.5
+chardet==4.0.0
+idna==2.10
+six==1.15.0
+urllib3==1.26.2
+wcwidth==0.2.5
